classwork final version

+++ 06-firebase/app/src/main/java/com/example/firebase_uno/CProducto.kt
@@ -22,7 +22,7 @@ class CProducto : AppCompatActivity() {
         val editTextPrecio = findViewById<EditText>(R.id.et_precioProd)
         val nuevoProducto = hashMapOf<String, Any>(
             "nombre" to editTextNombre.text.toString(),
-            "precio" to editTextPrecio.text.toString()
+            "precio" to editTextPrecio.text.toString().toDouble()
         )
         val db = Firebase.firestore
         val referencia = db.collection("producto")
